lab1/requirements.txt
torch>=2.2
torchvision>=0.17
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
pandas>=2.2
pillow>=10.0
kaggle>=1.6
notebook>=7.0
tqdm>=4.66

